backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-multipart
PyMuPDF
langchain
faiss-cpu
openai
langchain-community
python-dotenv
langchain-google-genai==1.0.10